Version 1.37
  * per-user-manager now accepts an optional listening FIFO open file descriptor, which it will use to listen for user-wide state change commands.
    system-control --user halt/normal/sysinit/&c. now send commands to this FIFO rather than sending signals to the wrong process.
  * The user-services@ service bundles now use fifo-listen and create the per-user-manager/ subdirectory.
  * The JAVA_HOME mechanism in the conversion subsystem is now unused by any service bundle.
  * setup-machine-id now writes /usr/local/etc/machine-id on FreeBSD/TrueOS systems.
  * static-networking configuration import now also sets up snort services.
  * There is now a Packages and Ports chapter in the Guide.
  * static-networking configuration import now handles ipv6_cpe_wanif and ipv6_activate_all_interfaces.
  * Added make-read-only-fs.
    It does not do much yet.
  * convert-systemd-units now recognizes CPUAffinity, ProtectHome, ProtectSystem, ReadWriteDirectories, ReadOnlyPaths, and InaccessiblePaths settings.
  * More common services: clickhouse-server, hue, udhcpc-log, minissdpd, rtkit-daemon, accounts-daemon, gdm3, speech-dispatcher, gdomap, blueman-mechanism
  * More BSD services: sysvipc
  * The user virtual terminal multiplexor and terminal emulator services now run under the aegides of their own dedicated user accounts, rather than daemon.
  * There is now a user-vt-realizer group to which users can be added to grant them realizer access to system-wide user virtual terminals.
  * Building now works even if one has a CDPATH.

+++ source/local-seqpacket-socket-listen.cpp
@@ -49,8 +49,8 @@ local_seqpacket_socket_listen (
 		popt::signed_number_definition uid_option('u', "uid", "number", "Specify the UID for the bound socket filename.", uid, 0);
 		popt::signed_number_definition gid_option('g', "gid", "number", "Specify the GID for the bound socket filename.", gid, 0);
 		popt::signed_number_definition mode_option('m', "mode", "number", "Specify the permissions for the bound socket filename.", mode, 0);
-		popt::string_definition user_option('\0', "user", "number", "Specify the user for the FIFO filename.", user);
-		popt::string_definition group_option('\0', "group", "number", "Specify the group for the FIFO filename.", group);
+		popt::string_definition user_option('\0', "user", "name", "Specify the user for the FIFO filename.", user);
+		popt::string_definition group_option('\0', "group", "name", "Specify the group for the FIFO filename.", group);
 		popt::definition * top_table[] = {
 			&upstart_compatibility_option,
 			&systemd_compatibility_option,
@@ -110,11 +110,14 @@ local_seqpacket_socket_listen (
 	if (0 > s) {
 exit_error:
 		const int error(errno);
-		std::fprintf(stderr, "%s: FATAL: %s\n", prog, std::strerror(error));
+		std::fprintf(stderr, "%s: FATAL: %s: %s\n", prog, listenpath, std::strerror(error));
 		throw EXIT_FAILURE;
 	}
-	unlink(listenpath);
-	if (0 > bind(s, reinterpret_cast<sockaddr *>(&addr), sizeof addr)) goto exit_error;
+	if (0 > bind(s, reinterpret_cast<sockaddr *>(&addr), sizeof addr)) {
+		if (EADDRINUSE != errno) goto exit_error;
+		unlink(listenpath);
+		if (0 > bind(s, reinterpret_cast<sockaddr *>(&addr), sizeof addr)) goto exit_error;
+	}
 	if (has_uid || has_gid) {
 		if (0 > chown(listenpath, uid, gid)) goto exit_error;
 	} else
